requirements.txt
asgiref==3.3.1
awesome-slugify==1.6.5
Django==3.1.6
django-annoying==0.10.6
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2021.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
Unidecode==0.4.21
whitenoise==5.2.0
